function_modules/image-to-text/requirements.txt
requests
redis
torch
pilgram
Pillow==6.2.2
zlib-state
transformers
